flask_rest_api/requirements.txt
flask
python-dotenv
flask_mysqldb
os